sweetalert added

+++ pages/index.js
@@ -3,6 +3,7 @@ import Link from 'next/link';
 import { useState } from 'react';
 import Container from '../components/container';
 import Global from '../components/global';
+import swal from 'sweetalert';
 
 const Index = ({todos}) =>{
 
@@ -44,28 +45,47 @@ const Index = ({todos}) =>{
                 </Container>
         )
 
-        async function onClickToDelete(ID){
-
-            // Call to delete method
-            const res = await fetch(Global.url + 'todo/' + ID, {
-                method: 'DELETE'
-            });
-        
-            const data = await res.json();
-        
-            //deleted ToDo
-            if (data.status == 'ok') {
-                //Call to API 
-                const res = await fetch(Global.url + 'todo');
-                const { todos } = await res.json()
-        
-                todoListModifier(todos);
-        
-            } else {
-                if (data.status == 'fail') {
-                    alert(data.error);
+        function onClickToDelete(ID){
+            //CONFIRMATION TO DELETE
+            swal({
+                title: "Are you sure?",
+                text: "Once deleted, you will not be able to recover this ToDo",
+                icon: "warning",
+                buttons: true,
+                dangerMode: true,
+              })
+              .then( async (willDelete) => {
+                if (willDelete) {
+                    /// WANT TO DELETE
+                        // Call to delete method
+                        const res = await fetch(Global.url + 'todo/' + ID, {
+                            method: 'DELETE'
+                        });
+                    
+                        const data = await res.json();
+                    
+                        //deleted ToDo
+                        if (data.status == 'ok') {
+                            //Call to API 
+                            const res = await fetch(Global.url + 'todo');
+                            const { todos } = await res.json()
+                    
+                            todoListModifier(todos);
+                            swal("Poof! The ToDo has been deleted!", {
+                                icon: "success",
+                            });
+                    
+                        } else {
+                            if (data.status == 'fail') {
+                                alert(data.error);
+                                swal("Error", data.error, "error");
+                            }
+                        }
+                    
+                } else {
+                    //nothing - no want to delete
                 }
-            }
+              }); 
         
         }
